+++ MealBytes/NetworkManager/FirebaseManager/FirebaseAuth/AuthError.swift
