+++ MealBytes/NetworkManager/FirebaseManager/FirebaseAuth/FirebaseAuth.swift
